muestro comision como float

+++ src/PagoAgilFrba/Rendicion/RendicionPago.cs
@@ -17,7 +17,7 @@ namespace PagoAgilFrba.Rendicion
     {
 
         RepoRendicion repo;
-        int totalRendicion;
+        float totalRendicion;
         List<Factura> facturas;
 
         public RendicionPago()
